refactor normalization into separate function
